fix and add types_of_variables
