+++ ydb/tests/datashard/lib/ya.make
@@ -2,6 +2,7 @@ PY3_LIBRARY()
 
 PY_SRCS(
     create_table.py
+    types_of_variables.py
 )
 
 END()
