+++ ydb/tests/stress/oltp_workload/workload/__init__.py
@@ -27,8 +27,7 @@ class WorkloadRunner:
     def _cleanup(self):
         print(f"Cleaning up {self.tables_prefix}...")
         deleted = self.client.remove_recursively(self.tables_prefix)
-        print(
-            f"Cleaning up {self.tables_prefix}... done, {deleted} tables deleted")
+        print(f"Cleaning up {self.tables_prefix}... done, {deleted} tables deleted")
 
     def run(self):
         stop = threading.Event()
